♻️ Refactor SubscriptionIO
