+++ client/types/MessageType.ts
@@ -1,0 +1,9 @@
+export type MessageType = {
+  id: number;
+  type: 'info' | 'error';
+  title: string;
+  message: string;
+  clickAction?: () => void | Promise<void>;
+  dismissed: boolean;
+  dismissDelay?: number;
+};
